# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recogno API",
    "description": "Backend for Recogno, a spaced-repetition trainer for DSA and system-design pattern recognition.\n\nThe drill loop is: `GET /drill/next` for a problem stripped of its answer, `POST /drill/submit` to grade a guess and advance its FSRS card, `GET /drill/due-count` for the reps-due badge.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /auth/login, /auth/register or the Google callback. Paste just the token — Swagger adds the \"Bearer \" prefix."
      }
    },
...
  },
  "paths": {
...
    "/drill/next": {
      "get": {
        "summary": "Get a problem to recognise",
        "tags": [
          "drill"
        ],
        "description": "Returns the statement and constraints only. The ground-truth pattern and the tell are withheld until the guess is submitted.\n\nPrefers an SRS card that is due, then a problem never seen, then the soonest-due card.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "problem": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "slug": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "statement": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "constraints": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "sourceUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "difficulty": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "easy",
                                "medium",
                                "hard"
                              ]
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "slug",
                        "title",
                        "statement",
                        "constraints",
                        "sourceUrl",
                        "difficulty"
                      ],
                      "additionalProperties": false
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "due",
                        "unseen",
                        "review-ahead"
                      ],
                      "description": "Which branch supplied this problem: a card that is due, a problem never seen, or the soonest-due card when nothing is due yet."
                    },
                    "dueAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "When the card fell due. Null if this problem is new to the user."
                    },
                    "patternOptions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "slug": {
                            "type": "string",
                            "example": "sliding-window"
                          },
                          "name": {
                            "type": "string",
                            "example": "Sliding Window"
                          },
                          "category": {
                            "type": "string",
                            "description": "Study grouping. Group the picker by this — there are ~80 patterns.",
                            "example": "Two Pointers"
                          }
                        },
                        "required": [
                          "id",
                          "slug",
                          "name",
                          "category"
                        ],
                        "additionalProperties": false
                      },
                      "description": "The full taxonomy to choose from. Reveals nothing about the answer."
                    }
                  },
                  "required": [
                    "problem",
                    "source",
                    "dueAt",
                    "patternOptions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Something the caller asked for does not exist or is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Something the caller asked for does not exist or is malformed."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}